pyproject.toml
[project]
name = "ai-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "instructor>=1.15.1",
    "litellm>=1.83.12",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.0.1",
    "rich>=14.3.4",
    "tavily-python>=0.7.23",
]
